LPD-60418 There are cases were the group might be deleted beforehand, so control this

+++ modules/apps/headless/headless-site/headless-site-test/src/testIntegration/java/com/liferay/headless/site/resource/v1_0/test/SiteResourceTest.java
@@ -75,7 +75,11 @@ public class SiteResourceTest extends BaseSiteResourceTestCase {
 		Collections.reverse(_sites);
 
 		for (Site site : _sites) {
-			_groupLocalService.deleteGroup(site.getId());
+			Group group = _groupLocalService.fetchGroup(site.getId());
+
+			if (group != null) {
+				_groupLocalService.deleteGroup(site.getId());
+			}
 		}
 
 		PrincipalThreadLocal.setName(_originalName);
